refactor: move enums to constants
