+++ lib/data/services/auth_service.dart
@@ -1,4 +1,4 @@
-import 'package:cribe/core/enums/storage_key.dart';
+import 'package:cribe/core/constants/storage_key.dart';
 import 'package:cribe/data/services/storage_service.dart';
 
 class AuthService {
