+++ lib/ui/auth/view_model/login_view_model.dart
@@ -1,6 +1,6 @@
-import 'package:cribe/core/enums/ui_state.dart';
-import 'package:cribe/core/services/validation_service.dart';
+import 'package:cribe/core/constants/ui_state.dart';
 import 'package:cribe/data/repositories/auth_repository.dart';
+import 'package:cribe/data/services/validation_service.dart';
 import 'package:flutter/foundation.dart';
 
 class LoginViewModel extends ChangeNotifier {
