+++ test/ui/home/widgets/home_screen_test.mocks.dart
@@ -6,7 +6,7 @@
 import 'dart:async' as _i5;
 import 'dart:ui' as _i6;
 
-import 'package:cribe/core/enums/ui_state.dart' as _i3;
+import 'package:cribe/core/constants/ui_state.dart' as _i3;
 import 'package:cribe/ui/home/view_model/home_view_model.dart' as _i2;
 import 'package:mockito/mockito.dart' as _i1;
 import 'package:mockito/src/dummies.dart' as _i4;
